requirements.txt
streamlit==1.25.0
qrcode==7.4.2
pillow==10.1.0
pandas==2.1.3
